requirements.txt
opencv-python
mediapipe
pyserial

Arduino
LED'S